server/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "coii-server"
version = "0.1.0"
description = "Open-source LLM experimentation platform"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.20.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-ulid>=2.0.0",
    "click>=8.1.0",
    "httpx>=0.27.0",
    "scipy>=1.13.0",
    "numpy>=1.26.0",
    "python-multipart>=0.0.9",
]

[project.scripts]
coii = "coii_server.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["coii_server"]
